+++ src/asr/index.ts
@@ -9,8 +9,8 @@ export {
   BaseASR,
   createASR,
   getASRProviders,
+  listen,
   registerASRProvider,
-  streamFrom,
 } from '@/asr/factory';
 
 export * from '@/asr/utils/index';